LPD-37459 Move the UndoRedo event listeners to the Page component
